refactor: words related recoils into words directory
