+++ src/hooks/use-reset-selected-chips.hook.ts
@@ -4,7 +4,7 @@ import { isFavoriteClickedState } from '@/recoil/favorites.state'
 import { selectedLanguageState } from '@/recoil/languages.state'
 import { selectedSemesterState } from '@/recoil/semesters.state'
 import { selectedCustomizedTagsState } from '@/recoil/tags.state'
-import { selectedCreatedDayState } from '@/recoil/created-date-tags.state'
+import { selectedCreatedDayState } from '@/recoil/words/created-date-tags.state'
 
 export const useResetSelectedChips = (code?: number) => {
   const setSelectedSemester = useSetRecoilState(selectedSemesterState)
