+++ src/recoil/words/favorites.state.ts
