+++ src/recoil/words/words.state.ts
